refacto: splitting into smaller functions and move them to new files
